feat(ui): standardize look for all app dialogs
Former-commit-id: e3bba31ca4f8f461efd9aaa73e354a25cc7e1577

+++ feature/mobile/search-expanded/src/main/kotlin/com/flixclusive/feature/mobile/searchExpanded/component/SearchSearchHistoryView.kt
@@ -22,7 +22,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.flixclusive.core.ui.common.CommonNoticeDialog
+import com.flixclusive.core.ui.common.dialog.TextAlertDialog
 import com.flixclusive.core.ui.common.util.onMediumEmphasis
 import com.flixclusive.feature.mobile.searchExpanded.SearchExpandedScreenViewModel
 import com.flixclusive.model.database.SearchHistory
@@ -82,7 +82,7 @@ internal fun SearchSearchHistoryView(
     }
 
     if (searchHistoryToDelete != null) {
-        CommonNoticeDialog(
+        TextAlertDialog(
             label = stringResource(id = UtilR.string.delete_search_history_item),
             description = stringResource(id = UtilR.string.delete_search_history_item_message),
             onConfirm = { viewModel.deleteSearchHistoryItem(searchHistoryToDelete!!) },
